fix(telegram): extract model_auth + fix misplaced test assertions
- Extract _HOST_AUTH_SETTING + is_host_authenticated into model_auth.py
  to bring model_picker.py under the 500-line budget (539 → 497).
- Move has_vendor_in_host assertions back into their correct test
  (test_has_host_and_has_vendor_in_host_guards) — they were misplaced
  in test_host_keyboard_hides_unauthenticated_providers where they
  fail because opencode_api_key is "" in that test's fixture.
