+++ backend/app/integrations/telegram/model_picker.py
@@ -1,8 +1,6 @@
 """Catalog-backed Telegram model picker helpers.
 
-The picker is a three-level walk: provider (host) → vendor → models.
-Hosts that serve a single vendor collapse to two levels (host → models),
-because forcing a single-button vendor screen is just an extra tap.
+Three-level walk: host → vendor → models (single-vendor hosts collapse to two levels).
 """
 
 from __future__ import annotations
@@ -14,11 +12,11 @@ from typing import Literal, Protocol
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.config import settings
 from app.core.providers.catalog import CATALOG_ETAG, MODEL_CATALOG, ModelEntry
 from app.core.providers.labels import host_label_from_slug, vendor_label_from_slug
 from app.crud.channel import get_or_create_telegram_conversation_full, get_user_id_for_external
 from app.crud.user_preferences import get_user_default_model_id
+from app.integrations.telegram.model_auth import is_host_authenticated
 from app.integrations.telegram.model_defaults import resolve_effective_model_id
 
 ModelCallbackAction = Literal["providers", "vendors", "list", "select", "set_default"]
@@ -41,39 +39,6 @@ _DEFAULT_BUTTON_TEXT = "⭐ Set as my default"
 _DEFAULT_ALREADY_SET_TEXT = "⭐ Already your default"
 # Inert callback used by the post-default badge so a second tap is a no-op.
 NOOP_CALLBACK = "mdl:noop"
-
-# Host → settings field that must be non-empty for the host to appear in
-# the picker (#370). Hosts that don't require a gateway API key
-# (``gemini-cli`` uses a local subprocess) are absent from this map and
-# always shown.
-_HOST_AUTH_SETTING: dict[str, str] = {
-    "agent-sdk": "claude_code_oauth_token",
-    "google-ai": "google_api_key",
-    "litellm": "openai_api_key",
-    "opencode-go": "opencode_api_key",
-    "xai": "xai_api_key",
-}
-
-
-def is_host_authenticated(host_slug: str) -> bool:
-    """Return whether the gateway-global config has the credentials for ``host_slug``.
-
-    Used by the picker to hide hosts whose credentials aren't configured
-    so the user doesn't pick a model that will immediately fail. Per-
-    workspace overrides aren't consulted here — the picker runs in the
-    chat surface (not the workspace), and a host that only some
-    workspaces have keys for is still better hidden than shown with a
-    failing default. When a workspace lands real per-user picker
-    filtering, that work threads workspace_root through this seam.
-
-    Hosts absent from :data:`_HOST_AUTH_SETTING` (e.g. ``gemini-cli``,
-    which runs a local subprocess rather than calling out to a managed
-    gateway) are always considered authenticated.
-    """
-    setting_name = _HOST_AUTH_SETTING.get(host_slug)
-    if setting_name is None:
-        return True
-    return bool(getattr(settings, setting_name, "") or "")
 
 
 class TelegramSenderLike(Protocol):
@@ -178,12 +143,7 @@ def build_host_keyboard() -> list[list[ModelButton]]:
 
 
 def build_vendor_keyboard(*, host: str) -> list[list[ModelButton]]:
-    """Build the vendor keyboard for a host that has multiple vendors.
-
-    For single-vendor hosts the caller should jump straight to
-    :func:`build_models_keyboard`; this function still works in that
-    case but produces a one-button screen.
-    """
+    """Build the vendor keyboard for a host with multiple vendors."""
     vendors = _host_to_vendors().get(host, {})
     rows: list[list[ModelButton]] = []
     current_row: list[ModelButton] = []
@@ -390,10 +350,8 @@ def _model_label(entry: ModelEntry, current_model_id: str) -> str:
 
 
 def _display_name_for_model(model_id: str) -> str:
-    for entry in MODEL_CATALOG:
-        if entry.id == model_id:
-            return entry.display_name
-    return model_id
+    entry = _entry_by_id(model_id)
+    return entry.display_name if entry else model_id
 
 
 def _catalog_index(entry: ModelEntry) -> int:
@@ -531,7 +489,6 @@ __all__ = [
     "get_model_picker_state",
     "has_host",
     "has_vendor_in_host",
-    "is_host_authenticated",
     "parse_model_callback_data",
     "picker_not_bound_message",
     "picker_stale_message",
